fix(categories): create components for category list and moved logic from HomePage to dedicated component
