+++ src/index.css
@@ -10,6 +10,7 @@
 :root {
   --orange: #ED6D33;
   --blue: #548AF3;
+  --light-blue: #3c7dff;
   --dark-blue: #548AF3;
   --purple: #C054F3;
   --light-purple: #d18df1;
