+++ src/pages/index.ts
@@ -1,3 +1,2 @@
 export {HomePage} from './HomePage/HomePage'
 export {UploadPage} from'./UploadPage/UploadPage'
-export {ErrorPage} from './Errorpage/ErrorPage'
