+++ src/utils/index.ts
@@ -1,4 +1,2 @@
 export {courses} from './categories'
 export {INSERT_API} from './endpoints'
-export {getRandomColor} from './GenerateRandomColor'
-export {lightenOrFade} from './lightenOrFade'
